# [feature] Update Kyo's foreach