--- a/kyo-core/shared/src/main/scala/kyo/Async.scala
+++ b/kyo-core/shared/src/main/scala/kyo/Async.scala
@@ -408,7 +408,7 @@ object Async extends AsyncPlatformSpecific:
         Frame
     ): Chunk[B] < (Abort[E] & Async & S) =
         if concurrency <= 1 then
-            Kyo.foreachIndexed(iterable.toSeq)(f)
+            Kyo.foreachIndexed(Chunk.from(iterable))(f)
         else
             iterable.size match
                 case 0 => Chunk.empty
@@ -421,8 +421,8 @@ object Async extends AsyncPlatformSpecific:
                 case size =>
                     isolate.capture { state =>
                         val groupSize = Math.ceil(size.toDouble / Math.max(1, concurrency)).toInt
-                        Fiber.foreachIndexed(iterable.grouped(groupSize).toSeq)((idx, group) =>
-                            Kyo.foreachIndexed(group.toSeq)((idx2, v) => isolate.isolate(state, f(idx + idx2, v)))
+                        Fiber.foreachIndexed(Chunk.from(iterable.grouped(groupSize)))((idx, group) =>
+                            Kyo.foreachIndexed(Chunk.from(group))((idx2, v) => isolate.isolate(state, f(idx + idx2, v)))
                         ).map(_.use(r => Kyo.foreach(r.flattenChunk)(isolate.restore)))
                     }