--- a/kyo-data/shared/src/main/scala/kyo/KArray.scala
+++ b/kyo-data/shared/src/main/scala/kyo/KArray.scala
@@ -119,6 +119,8 @@ object KArray:
           */
         def size: Int = self.length
 
+        inline def length: Int = size
+
         /** Checks if this KArray is empty.
           *
           * @return