--- a/kyo-direct/shared/src/test/scala/kyo/ShiftTest.scala
+++ b/kyo-direct/shared/src/test/scala/kyo/ShiftTest.scala
@@ -171,13 +171,13 @@ class ShiftMethodSupportTest extends AnyFreeSpec with Assertions:
                 assert(d.eval == Seq(1, 2, 3, 4))
             }
             "map - List" in {
-                val d: List[Int] < Any = defer:
+                val d: List[Int] < Any = direct:
                     xs.map(_.now).toList
 
                 assert(d.eval == List(1, 2, 3, 4))
             }
             "flatMap - List" in {
-                val d: List[Int] < Any = defer:
+                val d: List[Int] < Any = direct:
 
                     def f(i: Int) = List(1)
 
@@ -186,7 +186,7 @@ class ShiftMethodSupportTest extends AnyFreeSpec with Assertions:
                 assert(d.eval == List(1))
             }
             "collect - List" in {
-                val d: List[Int] < Any = defer:
+                val d: List[Int] < Any = direct:
                     xs.collect({
                         case i => i.now
                     }).toList
@@ -194,13 +194,13 @@ class ShiftMethodSupportTest extends AnyFreeSpec with Assertions:
                 assert(d.eval == List(1, 2, 3, 4))
             }
             "map - Vector" in {
-                val d: Vector[Int] < Any = defer:
+                val d: Vector[Int] < Any = direct:
                     xs.map(_.now).toVector
 
                 assert(d.eval == Vector(1, 2, 3, 4))
             }
             "flatMap - Vector" in {
-                val d: Vector[Int] < Any = defer:
+                val d: Vector[Int] < Any = direct:
 
                     def f(i: Int) = Vector(1)
 
@@ -209,7 +209,7 @@ class ShiftMethodSupportTest extends AnyFreeSpec with Assertions:
                 assert(d.eval == Vector(1))
             }
             "collect - Vector" in {
-                val d: Vector[Int] < Any = defer:
+                val d: Vector[Int] < Any = direct:
                     xs.collect({
                         case i => i.now
                     }).toVector