… kyo-kernel/shared/src/main/scala/kyo/Kyo.scala
… kyo-kernel/shared/src/main/scala/kyo/internal/KyoForeachSpecialized.scala
… kyo-kernel/shared/src/test/scala/kyo/KyoForeachTest.scala