### Problem
`foreach` only handle generic input, but not output the corresponding type
resolve #1020

### Solution
update to output the same collection type as input

### Notes

#### Limitiations:
I do think there are some new allocations here, especially around empty cases (`newBuilder` && `result` call can have some allocations). Maybe we can create our own `BuildFrom` that can cache the empty cases, or just create specialized function for each collection types