@HollandDM 2025-05-29
> There is one point, I think it's safer to use `source.iterableFactory.from(ch)` instead of source.newBuilder...

Yeah, this should be better for performance, as it's expose `empty` and `from` directly (which normally will be optimized on concrete classes), I'll update it later