@HollandDM 2025-05-31
@fwbrasil I took the `ForeachBench` and ran only first benchmark. Here is the result:

```
[info] Benchmark                 (size)   Mode  Cnt       Score      Error  Units
(current)
[info] ForeachBench.kyoForeach     1024  thrpt    3  168414.079 ± 3714.576  ops/s
[info] ForeachBench.kyoForeach    65536  thrpt    3    2566.369 ±   88.075  ops/s
[info] ForeachBench.kyoForeach   262144  thrpt    3     405.964 ±   11.555  ops/s
[info] ForeachBench.kyoForeach  1048576  thrpt    3     137.332 ±  259.822  ops/s
(this PR)
[info] ForeachBench.kyoForeach     1024  thrpt    3  111718.799 ± 3588.155  ops/s
[info] ForeachBench.kyoForeach    65536  thrpt    3    1441.030 ±   12.437  ops/s
[info] ForeachBench.kyoForeach   262144  thrpt    3     336.556 ±    8.316  ops/s
[info] ForeachBench.kyoForeach  1048576  thrpt    3      18.033 ±    1.110  ops/s
```
It seems that the overhead are quite big here.
Should we try specialized methods instead of generic like this?