@hearnadam 2025-06-22
@HollandDM I appreciate the work you put into exploring this and would love if it didn't go to waste.

What do you think about introducing `Kyo.foreach` alternatives for common iterables for now? This could defer the need to have a truly generic implementation.

One note, I don't think we should use the `KyoForeachSpecialized` trait, as it can make it harder for users to find methods.

If we need some indirectly, perhaps we can use inline methods in an object.