@HollandDM 2025-06-22
the trait was mainly used to separate the specialized code from the `Kyo.class` to make the file small.
I will try to come up with some implementation that doesn't blow out the code size. As of right now, it's quite big.