@ahoy-jon 2025-06-29
@HollandDM indeed, the CI fail on flaky.

What is your opinon on the implementation of `Kyo.foreach(List(1, 2, 3, ...))(_ + 1)` ? From what you have tested, is it better like that, or could it be better using `Kyo.foreach(Chunk.from ... )(_ + 1).map(_.toList)` ?