@HollandDM 2025-06-29
Append without the builder like this will result in a deep ‘Append’ node. ‘List.from’ then consume this  by convert chunk into indexed form, then do the ListBuffer builder and then accumulate into a list.
All of this is way more overhead compared to a single reverse, which has optimized code in List (one traversal only)