@ahoy-jon 2025-06-29
@HollandDM Indeed, ChunkBuilder is designed for concurrency and multi-shot usage, there's probably something worth exploring in that area.

Thanks again for the meticulous work 💪!

Note: This stuff is wild. I worked a bit on `Chunk`/`NonEmptyChunk` in ZIO back in 2020. There are multiple ways to implement a Chunk, plus Scala's boxing issues... honestly, this topic can drive anyone digging into it a little crazy 🤪