@HollandDM 2025-06-29
`Chunk` is way more complicated as they try to handle all the primitives themself (it will fail when dealing with types like `Int | Boolean`), and also aimed for the collection to be replacement for every other `Seq` in an ZIO app.
Kyo's `Chunk` is kind of like a `Array[Object]` with some special ops made optimized. It seems to trade simplicity for general usability. For example, in this PR I have to switch `foreachConcat` from old implementation to the new one with a (kind of) unconventional implementation (instead of concat right way, I keep those as nested Chunk, then flattenChunk in the end). But without it, the `foreachConcat` will be suffer a lot (my local benchmark shows that the old implementation is slower than both `List` and `Vector`)